longint puts

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -9,7 +9,7 @@
 
 void puts2(char *str)
 {
-	int n = 0;
+	long int n = 0;
 	while (str[n] != '\0')
 	{
 		_putchar(str[n]);
